% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ARG_MAZ}
\alias{ARG_MAZ}
\title{ARG_MAZ sapfluxnet site}
\format{
An sfn_data class object with the data and metadata for ARG_MAZ site
}
\usage{
ARG_MAZ
}
\description{
Example site for package usage demonstration based on ARG_MAZ
}
\examples{
data('ARG_MAZ', package = 'sapfluxnetr')
ARG_MAZ

}
\keyword{datasets}
