% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ARG_TRE}
\alias{ARG_TRE}
\title{ARG_TRE sapfluxnet site}
\format{
An sfn_data class object with the data and metadata for ARG_TRE site
}
\usage{
ARG_TRE
}
\description{
Example site for package usage demonstration based on ARG_TRE
}
\examples{
data('ARG_TRE', package = 'sapfluxnetr')
ARG_TRE

}
\keyword{datasets}
