% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{as_sfn_data_multi}
\alias{as_sfn_data_multi}
\title{as_sfn_data_multi helper}
\usage{
as_sfn_data_multi(x)
}
\arguments{
\item{x}{A list of sfn_data objects}
}
\value{
A sfn_data_multi object
}
\description{
Convert any list of sfn_data objects in a sfn_data_multi object
}
\examples{
sites_list <- list(ARG_TRE, ARG_MAZ)
sapfluxnetr:::as_sfn_data_multi(sites_list)

}
\keyword{internal}
