% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AUS_CAN_ST2_MIX}
\alias{AUS_CAN_ST2_MIX}
\title{AUS_CAN_ST2_MIX sapfluxnet site}
\format{An sfn_data class object with the data and metadata for AUS_CAN_ST2_MIX site}
\usage{
AUS_CAN_ST2_MIX
}
\description{
Example site for package usage demonstration based on AUS_CAN_ST2_MIX
}
\examples{
data('AUS_CAN_ST2_MIX', package = 'sapfluxnetr')
AUS_CAN_ST2_MIX

}
\keyword{datasets}
