% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BAR}
\alias{BAR}
\title{BAR sapfluxnet site}
\format{An sfn_data class object with the data and metadata for BAR site}
\usage{
BAR
}
\description{
Example site for package usage demonstration based on ARG_MAZ
}
\examples{
data('BAR', package = 'sapfluxnetr')
BAR

}
\keyword{datasets}
