% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BAZ}
\alias{BAZ}
\title{BAZ sapfluxnet site}
\format{An sfn_data class object with the data and metadata for BAZ site}
\usage{
BAZ
}
\description{
Example site for package usage demonstration based on AUS_CAN_ST2_MIX
}
\examples{
data('BAZ', package = 'sapfluxnetr')
BAZ

}
\keyword{datasets}
