% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{daily_metrics}
\alias{daily_metrics}
\title{Complete daily metrics for a site (or multi-site)}
\usage{
daily_metrics(sfn_data, solar = TRUE, pd_start = 3, pd_end = 5,
  md_start = 11, md_end = 13, probs = c(0.95, 0.99), ...)
}
\arguments{
\item{sfn_data}{\code{\link{sfn_data}} or \code{\link{sfn_data_multi}} object
to obtain the metrics from}

\item{solar}{Logical indicating if the solarTIMESTAMP must be used instead of
the site local TIMESTAMP. Default to TRUE (use solarTIMESTAMP).}

\item{pd_start}{Hour to start the predawn interval}

\item{pd_end}{Hour to end the predawn interval}

\item{md_start}{Hour to start the midday interval}

\item{md_end}{Hour to end the midday interval}

\item{probs}{numeric vector of probabilities for \code{\link[stats]{quantile}}}

\item{...}{optional arguments passed to \code{\link{sfn_metrics}}}
}
\value{
For \code{\link{sfn_data}} objects, a tibble with the metrics. For
  \code{\link{sfn_data_multi}} objects, a list of tibbles with the metrics
  for each site.
}
\description{
This function returns a complete daily summary for the site/s
}
\details{
\code{daily_metrics} is a wrapper for \code{\link{sfn_metrics}} with a set
of fixed arguments. It will always return the general metrics (for all data),
the predawn metrics (only data in the predawn period) and the midday metrics
(only data in the midday period).

\code{daily_metrics} returns the following statistics:
\itemize{
  \item{mean: daily mean of variable (tree or environmental variable). NAs
        are removed}
  \item{n: Number of measures used for the metrics (NAs included)}
  \item{coverage: Data coverage percentage (percentage of measures without
        NAs)}
  \item{q_XX: XX quantile value for the day}
  \item{max: Maximum value for the day}
  \item{max_time: Time when maximum value was reached}
  \item{min: Minimum value for the day}
  \item{min_time: Time when minimum value was reached}
  \item{centroid: Diurnal centroid value (hours passed until the half of
        the summed daily value was reached)}
}
}
\examples{
# data load
data('FOO', package = 'sapfluxnetr')

# default complete daily metrics
FOO_daily <- daily_metrics(FOO)

str(FOO_daily)
FOO_daily[['env']][['env_gen']]

# change the predawn and midday interval
FOO_int_daily <- daily_metrics(
  FOO,
  pd_start = 5, pd_end = 7, # predawn starting and ending hour
  md_start = 13, md_end = 15 # midday starting and ending hour
)

str(FOO_int_daily)

}
\seealso{
Other metrics: \code{\link{monthly_metrics}},
  \code{\link{sfn_metrics}}
}
