% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{describe_md_variable}
\alias{describe_md_variable}
\title{Detailed description of metadata variables}
\usage{
describe_md_variable(variable)
}
\arguments{
\item{variable}{A character with the name of the variable}
}
\value{
Nothing, prints information to console
}
\description{
\code{describe_md_variable} prints in console a detailed description for the
requested variable. Useful to know which values to filter or in which units
the variables are.
}
\examples{
# info about the method used to measure sapflow (pl_sens_meth)
describe_md_variable('pl_sens_meth')
}
