% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.accumulated_posix_aware}
\alias{.accumulated_posix_aware}
\title{accumulated function}
\usage{
.accumulated_posix_aware(variable, na.rm = FALSE)
}
\arguments{
\item{variable}{the variable as it is pased to the .fixed_metrics function}

\item{na.rm}{logical to pass to na.rm argument from sum}
}
\value{
the sum for any other variables, but the first for POSIXct variables
}
\description{
sum for summarise_all aware of POSIX objects
}
\examples{
sapfluxnetr:::.accumulated_posix_aware(c(50,20))
sapfluxnetr:::.accumulated_posix_aware(c(Sys.time(), Sys.time() -1))

}
\keyword{internal}
