% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{.assert_that_period_is_valid}
\alias{.assert_that_period_is_valid}
\title{test for character period}
\usage{
.assert_that_period_is_valid(period)
}
\arguments{
\item{period}{"frequency period"}
}
\value{
TRUE if is valid, informative error if not.
}
\description{
test for character period
}
\examples{
sapfluxnetr:::.assert_that_period_is_valid('1 day') # TRUE
# not run
# sapfluxnetr:::.assert_that_period_is_valid('1day') # error

}
\keyword{internal}
