% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.env_vars_names}
\alias{.env_vars_names}
\title{helper to return all environmental variable names}
\usage{
.env_vars_names()
}
\value{
a character vector with env vars names
}
\description{
helper to return all environmental variable names
}
\examples{
sapfluxnetr:::.env_vars_names()

}
\keyword{internal}
