% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{.fixed_metrics_funs}
\alias{.fixed_metrics_funs}
\title{helper function to generate the fixed metrics}
\usage{
.fixed_metrics_funs(probs, centroid)
}
\arguments{
\item{probs}{probs vector for quantile}

\item{centroid}{logical indicating if centroid calculation must be made.
(i.e. in monthly metrics, the centroid calculation is not needed)}
}
\description{
generates a call to dplyr::funs to capture the fixed metrics
}
