% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.metadata_architecture}
\alias{.metadata_architecture}
\title{Metadata variables architecture}
\usage{
.metadata_architecture()
}
\value{
a list with the metadata architecture
}
\description{
This function returns a nested list with the metadata, the variable and the
values, description , units and type for it use in describe_variable,
sfn_vars_to_filter and sfn_values_for
}
\examples{
sapfluxnetr:::.metadata_architecture()

}
\keyword{internal}
