% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{.nightly_daily_cf}
\alias{.nightly_daily_cf}
\title{Period custom function}
\usage{
.nightly_daily_cf(timestamp, night_data, int_start)
}
\arguments{
\item{timestamp}{TIMESTAMP}

\item{night_data}{as it comes inside of sfn_metrics}

\item{int_start}{interval start as supplied to sfn_metrics}
}
\value{
a vector of the same length as TIMESTAMP with the collapsed values.
}
\description{
nightly period custom function, to use with nightly_metrics
}
\details{
This function will be supplied as "period" argument to summarise_by_period
when daily nightly metrics are requested.
}
\keyword{internal}
