% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{.parse_period}
\alias{.parse_period}
\title{.parse_period}
\usage{
.parse_period(period)
}
\arguments{
\item{period}{Period in the "number period" format}
}
\value{
a list, with the freq value as numeric and the period value as
  character
}
\description{
Util to parse the period supplied to the function call to
create the collapsed timestamp
}
\examples{
sapfluxnetr:::.parse_period('1 day')

}
\keyword{internal}
