% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.sapflow_tidy}
\alias{.sapflow_tidy}
\title{.sapflow_tidy helper}
\usage{
.sapflow_tidy(data)
}
\arguments{
\item{data}{site sapflow metrics dataframe}
}
\description{
This helper is in charge of putting in shape the sapflow metrics, creating
the pl_code column.
}
\keyword{internal}
