% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.timezone_dic}
\alias{.timezone_dic}
\title{Timezones dictionary}
\usage{
.timezone_dic(tz)
}
\arguments{
\item{tz}{Character vector with the ISO code of the timezone as provided in
\code{env_time_zone} variable in \code{environmental_md}}
}
\value{
A character vector with the timezone code compatible with lubridate
  and as.POSIXct
}
\description{
Tranforms timezone ISO code to character vector compatible with lubridate and
POSIXct
}
\details{
GMT time zones are used, as they are day saving light time (DST) agnostic,
and in that way the DST can setted if the metadata says so. GMT are sign
exchanged to be compatible with ISO.
}
\keyword{internal}
