% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{filter_by_var}
\alias{filter_by_var}
\title{Filter the sites by metadata variable values}
\usage{
filter_by_var(variables, values, folder = ".", .use_cache = FALSE)
}
\arguments{
\item{variables}{A character vector indicating the variables to check}

\item{values}{A named list. Each element is named as the variable and contains
a character vector with the desired values for that variable}

\item{folder}{Route to the folder containing the data files (*.RData)}

\item{.use_cache}{Experimental, not implemented yet. Searches can be time
and resources consuming. Using a cache speed up the searches storing the
results in a cache folder and serving directly from the file.}
}
\value{
A character vector with the sites fullfilling the premises
}
\description{
\code{filter_by_var} function takes a vector of variables and a list of
possible values and list the sites that comply with the request
}
\details{
If \code{variables} length > 1, then only sites that match the values for
ALL variables declared are showed.

\code{variables} and \code{values} must be of the same length.
}
\examples{
# simple, want to know which sites are using the Heat Ratio method to measure
# the sap flow
filter_by_var('pl_sens_meth', 'HR', folder = 'Data')

# Both, Heat Ratio and Heat Dissipation
filter_by_var('pl_sens_meth',
              list(pl_sens_meth = c('HR', 'HD')),
              folder = 'Data')

# more complex, Heat Ratio method AND Mediterranean biome
filter_by_var(
  variables = c('pl_sens_meth', 'env_biome'),
  values = list(pl_sens_meth = 'HR',
                env_biome = 'Mediterranean'),
  folder = 'Data'
)

}
