% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FOO}
\alias{FOO}
\title{FOO sapfluxnet site}
\format{An sfn_data class object with the data and metadata for FOO site}
\usage{
FOO
}
\description{
Example site for package usage demonstration based on ARG_TRE
}
\examples{
data('FOO', package = 'sapfluxnetr')
FOO

}
\keyword{datasets}
