% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_timezone}
\alias{get_timezone}
\title{get the timezone of the site}
\usage{
get_timezone(sfn_data)
}
\arguments{
\item{sfn_data}{An sfn_data or sfn_data_multi object}
}
\value{
a character with the site timezone
}
\description{
Obtain the site timezone from a sfn_data/sfn_data_multi object
}
\examples{
# timezone of ARG_TRE site
get_timezone(ARG_TRE)

}
