% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_data_methods.R
\docType{methods}
\name{initialize,sfn_data_multi-method}
\alias{initialize,sfn_data_multi-method}
\title{Initialize method for sfn_data multi}
\usage{
\S4method{initialize}{sfn_data_multi}(.Object, ...)
}
\arguments{
\item{.Object}{sfn_data_multi object to create}

\item{...}{sfn_data elements}
}
\description{
Initialize an sfn_data_multi object
}
