% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_dplyr.R
\name{metrics_tidyfier}
\alias{metrics_tidyfier}
\title{Build a tidy data frame from the metrics results nested list}
\usage{
metrics_tidyfier(metrics_res, metadata, interval = "gen")
}
\arguments{
\item{metrics_res}{Nested list containing the metrics results as obtained
from \code{\link{metrics}}}

\item{metadata}{List containing the metadata nested list, as obtained from
\code{\link{read_sfn_metadata}}}

\item{interval}{Interval to return, it depends on the \code{metrics_res} and
can be \code{"gen"} for the general metrics, \code{"md"} for midday metrics,
\code{"pd"} for predawn metrics, \code{"night"} for night metrics or
\code{"day"} for diurnal metrics.}
}
\value{
a tibble with the following columns:
  \itemize{
    \item{TIMESTAMP: POSIXct vector with the date-time of the observation}
    \item{Tree}
  }
}
\description{
Transform the nested list of metrics in a tidy tibble where each observation
has its own row
}
\examples{
# not yet

}
