% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{min_max}
\alias{min_max}
\title{min max}
\usage{
min_max(x)
}
\arguments{
\item{x}{a numeric, POSIXct... (any accepted by \code{\link[base]{max}} and
\code{\link[base]{min}}) vector}
}
\value{
a two-element named vector, c(min = value, max = value)
}
\description{
wrapper for quicky return the max and the min value of a vector to use in
a dplyr pipe
}
\examples{
library(tidiverse)
# pipe example, not efficient
iris \%>\%
  pull(Sepal.Length) \%>\%
  min_max()

# the same, directly
min_max(iris$Sepal.Length)
}
