% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{monthly_metrics}
\alias{monthly_metrics}
\title{Complete monthly metrics for a site (or multi-site)}
\usage{
monthly_metrics(sfn_data, solar = TRUE, predawn = TRUE, pd_start = 3,
  pd_end = 5, midday = TRUE, md_start = 11, md_end = 13,
  probs = c(0.95, 0.99), na.rm = FALSE, ...)
}
\arguments{
\item{sfn_data}{\code{\link{sfn_data}} or \code{\link{sfn_data_multi}} object
to obtain the metrics from}

\item{solar}{Logical indicating if the solarTIMESTAMP must be used instead of
the site local TIMESTAMP. Default to TRUE (use solarTIMESTAMP).}

\item{predawn}{Logical indicating if metrics for predawn interval must be
also returned.}

\item{pd_start}{Hour to start the predawn interval}

\item{pd_end}{Hour to end the predawn interval}

\item{midday}{Logical indicating if metrics for midday interval must be also
returned.}

\item{md_start}{Hour to start the midday interval}

\item{md_end}{Hour to end the midday interval}

\item{probs}{numeric vector of probabilities for \code{\link[stats]{quantile}}}

\item{...}{optional arguments passed to \code{\link{sfn_metrics}}}
}
\value{
For \code{\link{sfn_data}} objects, a tibble with the metrics. For
  \code{\link{sfn_data_multi}} objects, a list of tibbles with the metrics
  for each site.
}
\description{
This function returns a complete summary of the site/s
}
\seealso{
Other metrics: \code{\link{daily_metrics}},
  \code{\link{sfn_metrics}}
}
