% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{period_summaries}
\alias{period_summaries}
\title{metrics shared statistics}
\usage{
period_summaries(sfn_data, period, .funs, ...)
}
\arguments{
\item{sfn_data}{an sfn_data object}

\item{period}{tibbletime::collapse_by period}

\item{.funs}{summarise_all funs}

\item{...}{optional arguments for tibbletime::collapse_by function}
}
\description{
helper for sfn_metrics
}
\details{
This helper pipes the different chain of commands to perform the period
summaries, shared by general, predawn and midday metrics
}
