% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{read_sfn_data}
\alias{read_sfn_data}
\title{Read sfn_data from disk}
\usage{
read_sfn_data(site_codes, folder = ".")
}
\arguments{
\item{site_codes}{A character vector with the site code/s}

\item{folder}{Route to the folder containing the \code{.RData} file. Default
to working directory.}
}
\value{
If \code{site_codes} is a vector of length 1, an sfn_data object with
  the selected site data. If \code{site_codes} is a vector of length > 1, then
  a sfn_data_multi object containing all selected sites.
}
\description{
Given a site code and a route, \code{read_sfn_data} will return the selected
sfn_data object
}
\examples{
\dontrun{
  # Let's access the plant level data, in the 'RData/plant' folder
  # (change accordingly if your data is in another location or
  # you want the leaf or sapwood unit level)
  path_to_data <- file.path('RData', 'plant')
  
  # now we read a single site
  ARG_TRE <- read_sfn_data('ARG_TRE', path_to_data)
  ARG_TRE
  
  # or we can read multiple sites at once
  multi_sfn <- read_sfn_data(
    c('ARG_TRE', 'ARG_MAZ', 'AUS_CAN_ST2_MIX'), path_to_data
  )
  multi_sfn
}

}
