% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{read_sfn_metadata}
\alias{read_sfn_metadata}
\title{Read and combine all metadata}
\usage{
read_sfn_metadata(folder = ".", .write_cache = FALSE)
}
\arguments{
\item{folder}{Route to the folder containing the data. Default to working
directory}

\item{.write_cache}{Logical indicating if a cached copy of the metadata must
be written in \code{folder}.}
}
\value{
A list of tibbles with the five metadata classes (site, stand,
  species, plant and environmental)
}
\description{
Read metadata from all sites in folder and write it to disk to cache the
info for easy and fast access
}
\details{
Load all data in memory to collect metadata info can be resource limiting.
For easy and quick access to metadata, this function stores an .RData file
in the specified folder along the data with all the metadata preloaded. Also
it return it as an object to use in filtering and selecting sites.
}
\examples{
\dontrun{
# load the metadata for the first time, it can be a minute ;)
sites_metadata <- read_metadata(folder = 'Data', .write.cache = TRUE)

# a cached copy must have been written to "folder"
file.exists('Data/.metadata_cache.RData') # TRUE

# inspect the metadata
sites_metadata
}

}
