% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_data_classes.R
\docType{class}
\name{sfn_data_multi-class}
\alias{sfn_data_multi-class}
\alias{sfn_data_multi}
\title{S4 class for sapfluxnet multi-site data}
\description{
Multi sfn data class, derived from list
}
\details{
This class inherits from \code{list}, but modified to contain sfn_data objects
as elements. This will allow to work with several sites at the same time
obtaining results for all of them combined or individually as elements of
the resulting list (with \code{lapply} or \code{purrr::map})
}
