% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_data_methods.R
\name{sfn_data_multi_validity}
\alias{sfn_data_multi_validity}
\title{Validity method for sfn_data_multi class}
\description{
Validation checks for generating sfn_data_multi class objects
}
\details{
This method is used internally to ensure the correctnes of the sfn_data_multi
object. Basically ensures that the object returned is a list of sfn_data
class objects
}
