% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_data_classes.R
\docType{data}
\name{sfn_data_R6}
\alias{sfn_data_R6}
\title{R6 clas for sapfluxnet site data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sfn_data_R6
}
\description{
R6 clas for sapfluxnet site data
}
\keyword{datasets}
