% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_data_methods.R
\name{sfn_data_validity}
\alias{sfn_data_validity}
\title{Validity method for sfn_data class}
\description{
Validation checks for generating sfn_data class objects
}
\details{
This method is used internally when creating and/or modifing sfn_data class
objects to ensure that the object returned is correct in terms of content
classes and dimensions (i.e. sapflow data and environmental data has the
same lenght)
}
