% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sfn_filter}
\alias{sfn_filter}
\title{Filtering function for sfn_data objects}
\usage{
sfn_filter(sfn_data, ..., solar = FALSE)
}
\arguments{
\item{sfn_data}{\code{sfn_data} or \code{sfn_data_multi} object to subset}

\item{...}{filter expressions to subset the data, as in
\code{\link[dplyr]{filter}}}

\item{solar}{Logical indicating if solar timestamp must used to subset}
}
\value{
For \code{sfn_data} objects, a filtered \code{sfn_data} or NULL if
  no data meet the criteria. For \code{sfn_data_multi} another
  \code{sfn_data_multi} with the sites filtered, and an empty
  \code{sfn_data_multi} if any sites met the criteria
}
\description{
Filter an \code{\link{sfn_data}} object by matching values
}
\details{
If \code{sfn_data} is an sfn_data_multi object, all sites are subsetted in
the same way. This is a convenient way of filtering sites that are
concurrent in time (i.e. selecting sites for a metadata variable)
}
\examples{
library(dplyr)
library(lubridate)

# data
data('FOO', package = 'sapfluxnetr')

# by timestamp
foo_timestamp <- get_timestmap(FOO)

foo_timestamp_trimmed <- foo_timestamp[1:100]

sfn_filter(
  FOO,
  TIMESTAMP \%in\% foo_timestamp_trimmed
)

# by wind speed value
ws_threshold <- 25

sfn_filter(
  FOO,
  ws <= ws_threshold
)

## multi
data('BAR', package = 'sapfluxnetr')
multi_sfn <- sfn_data_multi(FOO, BAR)

# by timestamp
sfn_filter(
  multi_sfn,
  between(day(TIMESTAMP), 18, 22)
)

# by wind speed value
sfn_filter(
  multi_sfn,
  ws <= ws_threshold
)

}
