% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_data_methods.R
\name{sfn_get_methods}
\alias{sfn_get_methods}
\title{sfn_data get methods}
\arguments{
\item{object}{Object of class sfn_data from which data is retrieved}

\item{solar}{Logical indicating if the timestamp to return in the get_sapf,
get_env, get_sapf_flags and get_env_flags methods}
}
\description{
Methods to get the data and metadata from the sfn_data class slots
}
\details{
\code{get_sapf} and \code{get_env} methods retrieve sapflow or environmental
data and timestamp to create a functional dataset to work with.

\code{get_sapf_flags} and \code{get_env_flags} methods retrieve sapflow or
environmental flags also with the timestamp.

\code{get_timestamp} method retrieve only the timestamp as POSIXct vector.

\code{get_si_code} method retrieve a character vector with length(timestamp)
containing the site code.

\code{get_site_md}, \code{get_stand_md}, \code{get_species_md},
\code{get_plant_md} and \code{get_env_md} methods retrieve the corresponding
metadata.
}
