% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfn_metadata_ex}
\alias{sfn_metadata_ex}
\title{sfn_metadata cache file for example data (ARG_MAZ, ARG_TRE and AUS_CAN_ST2_MIX)}
\format{A list with five elements, each of one a metadata type.}
\usage{
sfn_metadata_ex
}
\description{
Example metadata cache file content for package usage demonstration
}
\examples{
data('sfn_metadata_ex', package = 'sapfluxnetr')
sfn_metadata_ex

}
\keyword{datasets}
