% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{sfn_plot}
\alias{sfn_plot}
\title{plot method for sfn_data class}
\usage{
sfn_plot(sfn_data, type = c("sapf", "env", "ta", "rh", "vpd", "ppfd_in",
  "netrad", "sw_in", "ext_rad", "ws", "precip", "swc_shallow", "swc_deep"),
  solar = TRUE)
}
\arguments{
\item{sfn_data}{sfn_data object to plot}

\item{type}{Character indicating which data to plot. See Type section for
detailed information about the available values}

\item{solar}{Logical indicating if the solar timestamp must be used instead
of the site timestamp}
}
\value{
A ggplot object that can be called to see the plot
}
\description{
Plot the desired data from a site object
}
\section{ggplot plotting system}{

  \code{\link{plot}} is a base R function which uses the base R plotting system
  to show the plot. We prefer the ggplot plotting system, which allow for
  storing the plots in objects and can be subjected to further modifications.
  This allow the package users to generate rather simple plots that can be
  fine tuned afterwards to the user taste. Generating a \code{\link{plot}}
  method for the \code{sfn_data} class returning a ggplot object is not
  desired (it change the way plot works and can be misleading about the plot
  general usage). So, instead, we offer this function, \code{sfn_plot}.
}

\examples{
# data
data('FOO', package = 'sapfluxnetr')

# plotting directly
sfn_plot(FOO, type = 'sapf', solar = FALSE)

# saving and modifying:
env_plot <- sfn_plot(FOO, type = 'env', solar = FALSE)
env_plot + title('Environmental variables facet plot')

}
