% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{sfn_sites_in_folder}
\alias{sfn_sites_in_folder}
\title{list available sites in a db folder}
\usage{
sfn_sites_in_folder(folder = ".")
}
\arguments{
\item{folder}{Character vector of length 1 indicating the route to the
db folder}
}
\description{
Retrieves the site codes in the specified folder
}
\details{
If folder
}
\examples{
\dontrun{
# list the sites available in Data folder
sfn_sites_in_folder('Data/')
}

}
