% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sfn_subset}
\alias{sfn_subset}
\title{Subsetting function for sfn_data objects}
\usage{
sfn_subset(sfn_data, ..., solar = FALSE)
}
\arguments{
\item{sfn_data}{\code{sfn_data} or \code{sfn_data_multi} object to subset}

\item{...}{filter expressions to subset the data}

\item{solar}{Logical indicating if solar timestamp must used to subset}
}
\value{
An \code{sfn_data} or \code{sfn_data_multi} object subsetted as per
  \code{...}
}
\description{
Subset an \code{\link{sfn_data}} object by matching values
}
\details{
If \code{sfn_data} is an sfn_data_multi object, all sites are subsetted in
the same way.
}
\examples{
library(dplyr)
library(lubridate)

# data
data('FOO', package = 'sapfluxnetr')

# by timestamp
foo_timestamp <- get_timestmap(FOO)

foo_timestamp_trimmed <- foo_timestamp[1:100]

sfn_subset(
  FOO,
  TIMESTAMP \%in\% foo_timestamp_trimmed
)

# by wind speed value
ws_threshold <- 25

sfn_subset(
  FOO,
  ws <= ws_threshold
)

## multi
data('BAR', package = 'sapfluxnetr')
multi_sfn <- sfn_data_multi(FOO, BAR)

# by timestamp
sfn_subset(
  multi_sfn,
  between(day(TIMESTAMP), 18, 22)
)

# by wind speed value
sfn_subset(
  multi_sfn,
  ws <= ws_threshold
)

}
