% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sfn_values_for}
\alias{sfn_values_for}
\title{List all possible values for a filtering metadata variable}
\usage{
sfn_values_for(var = "pl_sens_meth")
}
\arguments{
\item{var}{Character indicating the metadata variable}
}
\value{
A character vector with the accepted values for the desired variable
}
\description{
\code{sfn_values_for} function returns a vector with the accepted values for
the required metadata variable, and can be used to filter the sites with
\code{\link{filter_by_var}}
}
\examples{
# accepted values for plant sensor method (pl_sens_meth)
sfn_values_for('pl_sens_meth')

}
