% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sfn_vars_to_filter}
\alias{sfn_vars_to_filter}
\title{List all variables that can be used to filter sites}
\usage{
sfn_vars_to_filter()
}
\value{
A list with five elements, \code{site_md}, \code{stand_md},
  \code{species_md}, \code{plant_md} and \code{env_md}
}
\description{
\code{sfn_vars_to_filter()} returns a list with the variables for each
kind of metadata that can be used to select and filter sites
}
\examples{
# all variables
sfn_vars_to_filter()

# by some metadata
sfn_vars_to_filter()$site_md

}
