% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_data_methods.R
\name{show,sfn_data-method}
\alias{show,sfn_data-method}
\title{Show method for sfn_data}
\usage{
\S4method{show}{sfn_data}(object)
}
\arguments{
\item{object}{sfn_data object to show}
}
\description{
print a summary for sfn_data objects
}
