% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_data_methods.R
\name{show,sfn_data_multi-method}
\alias{show,sfn_data_multi-method}
\title{Show method for sfn_data_multi}
\usage{
\S4method{show}{sfn_data_multi}(object)
}
\arguments{
\item{object}{sfn_data_multi object to show}
}
\description{
print a summary for sfn_data_multi objects
}
