% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{filter_by_var}
\alias{filter_by_var}
\title{Filter the sites by metadata variable values}
\usage{
filter_by_var(..., folder = ".", join = c("and", "or"),
  .use_cache = FALSE)
}
\arguments{
\item{...}{Logical expressions for the metadata variables, as in
\code{\link[dplyr]{filter}}.}

\item{folder}{Route to the folder containing the data files (*.RData)}

\item{join}{Character indicating how to filter the sites, see details.}

\item{.use_cache}{Experimental, not implemented yet. Searches can be time
and resources consuming. Using a cache speed up the searches storing the
results in a cache folder and serving directly from the file.}
}
\value{
A character vector with the sites fullfilling the premises
}
\description{
\code{filter_by_var} function takes logical expressions for the metadata
variables (i.e. \code{pl_sens_meth == 'HR'}), and list the sites that meet
the expressions.
}
\details{
\code{join} argument indicates how sites must be filtered between metadata
classes. \code{'and'} indicates only sites meeting all conditions for all
metadata classes are returned. \code{'or'}cindicates all sites meeting any
condition between classes are returned. For two or more filtes of the same
metadata class, they are combined as 'and'.
}
\examples{
# simple, want to know which sites are using the Heat Ratio method to measure
# the sap flow
filter_by_var(pl_sens_meth == 'HR', folder = 'Data')

# Both, Heat Ratio and Heat Dissipation
filter_by_var(pl_sens_meth \%in\% c('HR', 'HD'),
              folder = 'Data')

# more complex, Heat Ratio method AND Mediterranean biome
filter_by_var(
  pl_sens_meth == 'HR',
  si_biome == 'Mediterranean',
  folder = 'Data',
  join = 'and' # default
)

# join = 'or' returns sites that meet any condition
filter_by_var(
  pl_sens_meth == 'HR',
  si_biome == 'Mediterranean',
  folder = 'Data',
  join = 'or'
)

}
