% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\alias{daily_metrics}
\alias{monthly_metrics}
\alias{nighttime_metrics}
\title{Complete metrics wrappers}
\usage{
daily_metrics(sfn_data, solar = TRUE, pd_start = 3, pd_end = 5,
  md_start = 11, md_end = 13, probs = c(0.95, 0.99), ...)

monthly_metrics(sfn_data, solar = TRUE, pd_start = 3, pd_end = 5,
  md_start = 11, md_end = 13, probs = c(0.95, 0.99), ...)

nighttime_metrics(sfn_data, period = c("daily", "monthly"), solar = TRUE,
  night_start = 20, night_end = 6, probs = c(0.95, 0.99), ...)
}
\arguments{
\item{sfn_data}{\code{\link{sfn_data}} or \code{\link{sfn_data_multi}} object
to obtain the metrics from}

\item{solar}{Logical indicating if the solarTIMESTAMP must be used instead of
the site local TIMESTAMP. Default to TRUE (use solarTIMESTAMP).}

\item{pd_start}{Hour to start the predawn interval}

\item{pd_end}{Hour to end the predawn interval}

\item{md_start}{Hour to start the midday interval}

\item{md_end}{Hour to end the midday interval}

\item{probs}{numeric vector of probabilities for \code{\link[stats]{quantile}}}

\item{...}{Not implemented}

\item{period}{Time period to aggregate data by. See period section for an explanation
about the periods ('daily', 'monthly', 'yearly', ...)}

\item{night_start}{Hour to start the night interval}

\item{night_end}{Hour to end the night interval}
}
\value{
For \code{\link{sfn_data}} objects, a tibble with the metrics. For
  \code{\link{sfn_data_multi}} objects, a list of tibbles with the metrics
  for each site.
}
\description{
This set of functions returns a complete set of statistics for a site or
several sites (using \code{\link{sfn_data_multi}})
}
\details{
\code{*_metrics} functions are wrappers for \code{\link{sfn_metrics}} with a
set of fixed arguments.

\code{*_metrics} functions return all or some of the following statistics:
\itemize{
  \item{mean: mean of variable (tree or environmental variable) for the
        given period. NAs are removed}
  \item{n: Number of measures in the give period used for the metrics (NAs included)}
  \item{coverage: Data coverage percentage (percentage of measures without
        NAs)}
  \item{q_XX: XX quantile value for the period}
  \item{max: Maximum value for the period}
  \item{max_time: Time when maximum value was reached}
  \item{min: Minimum value for the period}
  \item{min_time: Time when minimum value was reached}
  \item{centroid: Diurnal centroid value (hours passed until the half of
        the summed daily value was reached). Only returned when period
        is 'daily'}
}
}
\section{daily_metrics}{

\code{daily_metrics} will always return the general metrics (for all data),
the predawn metrics (only data in the predawn period) and the midday metrics
(only data in the midday period).
}

\section{monthly_metrics}{

\code{monthly_metrics} will always return the general metrics (for all data),
the predawn metrics (only data in the predawn period) and the midday metrics
(only data in the midday period).
}

\section{nighttime_metrics}{

\code{nighttime_metrics} will always return the metrics for day and night
periods, summarised daily or monthly

Night for daily period starts in DOY x and ends in DOY x+1 (i.e. if
\code{night_start = 20, night_end = 6} values for 2018-03-28 20:00:00 means
values for the night starting at 2018-03-28 20:00:00 and ending at
2018-03-29 06:00:00).

Night for monthly period summarises all night in the month (including that
starting the last day of the month and ending the first day of the next).
}

\examples{
## daily_metrics
# data load
data('FOO', package = 'sapfluxnetr')

# default complete daily metrics
FOO_daily <- daily_metrics(FOO)

str(FOO_daily)
FOO_daily[['env']][['env_gen']]

# change the predawn and midday interval
FOO_int_daily <- daily_metrics(
  FOO,
  pd_start = 5, pd_end = 7, # predawn starting and ending hour
  md_start = 13, md_end = 15 # midday starting and ending hour
)

str(FOO_int_daily)

## monthly_metrics
# data load
data('BAR', package = 'sapfluxnetr')

# default complete daily metrics
BAR_monthly <- monthly_metrics(BAR)

str(BAR_monthly)
BAR_monthly[['env']][['env_gen']]

# change the predawn and midday interval
BAR_int_monthly <- monthly_metrics(
  BAR,
  pd_start = 5, pd_end = 7, # predawn starting and ending hour
  md_start = 13, md_end = 15 # midday starting and ending hour
)

## nighttime_metrics
# data load
data('BAZ', package = 'sapfluxnetr')

# default complete daily metrics
BAZ_monthly <- nighttime_metrics(BAR, period = 'monthly')

str(BAZ_monthly)
BAZ_monthly[['env']][['env_day']]
BAZ_monthly[['env']][['env_night']]

# change the night interval
BAZ_monthly_short <- nighttime_metrics(
  BAR,
  night_start = 21, night_end = 4 # night starting and ending hour
)

BAZ_monthly_short[['env']][['env_night']]

}
\seealso{
Other metrics: \code{\link{sfn_metrics}}
}
