% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{read_sfn_data}
\alias{read_sfn_data}
\title{Read sfn_data from disk}
\usage{
read_sfn_data(site_code, folder = ".")
}
\arguments{
\item{site_code}{A character vector with the site code/s}

\item{folder}{Route to the folder containing the \code{.RData} file. Default
to working directory.}
}
\value{
If \code{site_code} is a vector of length 1, an sfn_data object with
  the selected site data. If \code{site_code} is a vector of length > 1, then
  a sfn_data_multi object containing all selected sites.
}
\description{
Given a site code and a route, \code{read_sfn_data} will return the selected
sfn_data object
}
\examples{
# load one single site (it suposes a folder called Data exists and contains
# a file called FOO.RData)
\dontrun{
FOO <- read_sfn_data('FOO', 'Data')

# load several sites
multi_sfn <- read_sfn_data(c('FOO', 'BAR', 'BAZ'), 'Data')
}

}
