% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_data_generics.R
\name{sfn_get_generics}
\alias{sfn_get_generics}
\alias{get_sapf_data}
\alias{get_env_data}
\alias{get_sapf_flags}
\alias{get_env_flags}
\alias{get_timestamp}
\alias{get_solar_timestamp}
\alias{get_si_code}
\alias{get_site_md}
\alias{get_stand_md}
\alias{get_species_md}
\alias{get_plant_md}
\alias{get_env_md}
\title{sfn_data custom get generics}
\usage{
get_sapf_data(object, ...)

get_env_data(object, ...)

get_sapf_flags(object, ...)

get_env_flags(object, ...)

get_timestamp(object, ...)

get_solar_timestamp(object, ...)

get_si_code(object, ...)

get_site_md(object, ...)

get_stand_md(object, ...)

get_species_md(object, ...)

get_plant_md(object, ...)

get_env_md(object, ...)
}
\arguments{
\item{object}{Object to get data from}

\item{...}{Further arguments to pass to the corresponding get method}
}
\description{
Generics for getting the info in the slots of SfnData
}
\details{
see \code{\link{sfn_get_methods}} for detailed info about using the get
methods in \code{sfn_data} class objects and
\code{\link{sfn_multi_get_methods}} for detailed info about using the get
methods in \code{sfn_data_multi} class objects.
}
