% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sfn_metrics}
\alias{sfn_metrics}
\title{Metrics summary function}
\usage{
sfn_metrics(sfn_data, period, .funs, solar, predawn, pd_start, pd_end, midday,
  md_start, md_end, ...)
}
\arguments{
\item{sfn_data}{\code{\link{sfn_data}} or \code{\link{sfn_data_multi}} object
to obtain the metrics from}

\item{period}{Time period to aggregate data by. See period section for an explanation
about the periods ('daily', 'monthly', 'yearly', ...)}

\item{.funs}{List of function calls generated by \code{\link[dplyr]{funs}}, or
a character vector of functions means, or simply a function. See .funs section
for more info about supplying custom functions.}

\item{solar}{Logical indicating if the solarTIMESTAMP must be used instead of
the site local TIMESTAMP. Default to TRUE (use solarTIMESTAMP).}

\item{predawn}{Logical indicating if metrics for predawn interval must be
also returned.}

\item{pd_start}{Hour to start the predawn interval}

\item{pd_end}{Hour to end the predawn interval}

\item{midday}{Logical indicating if metrics for midday interval must be also
returned.}

\item{md_start}{Hour to start the midday interval}

\item{md_end}{Hour to end the midday interval}

\item{...}{optional arguments to pass to methods used
(i.e. tibbletime::collapse_index or summarise funs extra arguments)}

\item{nighttime}{Experimental, not implemented (stats for night time)}
}
\value{
For \code{\link{sfn_data}} objects, a tibble with the metrics. For
  \code{\link{sfn_data_multi}} objects, a list of tibbles with the metrics
  for each site.
}
\description{
Generate daily or above metrics from a site data for the period indicated
}
\section{period}{

\code{period} argument is piped to \code{tibbletime::collapse_index} function.
See \code{\link[tibbletime]{collapse_index}} for a detailed explanation but in
short:
\itemize{
  \item{\emph{frequency period} format: "1 day", "7 day", "1 month", "1 year"}
  \item{\emph{shorthand} format: "hourly", "daily", "monthly", "yearly"}
  \item{\emph{custom} format: a vector of dates to use as custom and more flexible boundaries}
}
}

\section{.funs}{

\code{.funs} argument uses the same method as the \code{.funs} argument in the
\code{\link[dplyr]{summarise_all}} function of \code{dplyr} package. Basically
it accepts a list of function calls generated by funs(), or a character vector
of function names, or simply a function. If you want to pass on a custom
function you can specify it here.
}

\examples{
## sfn_data
data('FOO', pkg = 'sapfluxnetr')
sfn_metrics(FOO)

## sfn_data_multi
data('sfn_multi', pkg = 'sapfluxnetr')
sfn_metrics(FOO)

}
\seealso{
Other metrics: \code{\link{daily_metrics}},
  \code{\link{monthly_metrics}}
}
