% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_data_generics.R
\name{sfn_replacement_generics}
\alias{sfn_replacement_generics}
\alias{get_sapf_data<-}
\alias{get_env_data<-}
\alias{get_sapf_flags<-}
\alias{get_env_flags<-}
\alias{get_timestamp<-}
\alias{get_solar_timestamp<-}
\alias{get_si_code<-}
\alias{get_site_md<-}
\alias{get_stand_md<-}
\alias{get_species_md<-}
\alias{get_plant_md<-}
\alias{get_env_md<-}
\title{sfn_data replacement generics}
\usage{
get_sapf_data(object) <- value

get_env_data(object) <- value

get_sapf_flags(object) <- value

get_env_flags(object) <- value

get_timestamp(object) <- value

get_solar_timestamp(object) <- value

get_si_code(object) <- value

get_site_md(object) <- value

get_stand_md(object) <- value

get_species_md(object) <- value

get_plant_md(object) <- value

get_env_md(object) <- value
}
\arguments{
\item{object}{Object to replace}

\item{value}{Object to replace with}
}
\description{
Generic functions for replacement functions for sfn_data
}
\details{
see \code{\link{sfn_replacement_methods}} for more info about using the
replacement methods in sfn_data objects
}
