% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{sfn_sites_in_folder}
\alias{sfn_sites_in_folder}
\title{list available sites in a db folder}
\usage{
sfn_sites_in_folder(folder = ".")
}
\arguments{
\item{folder}{Character vector of length 1 indicating the route to the
db folder}
}
\value{
A character vector with the site codes present in the folder, an
  error if the folder is not valid or does not contain any site data file.
}
\description{
Retrieves the site codes in the specified folder
}
\details{
If folder
}
\examples{
\dontrun{
# Let's access the plant level data, in the 'RData/plant' folder
# (change accordingly if your data is in another location or
# you want the leaf or sapwood unit level)
path_to_data <- file.path('RData', 'plant')

# lets see the sites
sfn_sites_in_folder(path_to_data)
}

}
